Changes to mainloop, added widget.
Added a few FIFO widget that allows you to monitor a fifo for incoming
data.  As data comes in it is printed in the info bar and fades over
time.  This can be used as a notification daemon for different apps.

Changed the mainloop so that plugins can register file descriptors
(sockets, files, fifos) so that they don't need to spawn new threads
(which is bad in python) and moved the X11 events to use this system.
This way only when a new x11 event comes in does the callback get run.

Fixed the volume widget to not call an explicit redraw to the taskbar,
but rather just tell the taskbar it needs to update.  This will work
fine as when mouse input is given it automatically updates.